+++ packages/react-reconciler/src/ReactFiberCommitWork.js
@@ -158,6 +158,7 @@ import {
   supportsHydration,
   supportsResources,
   supportsSingletons,
+  supportsViewTransition,
   clearSuspenseBoundary,
   clearSuspenseBoundaryFromContainer,
   createContainerChildSet,
@@ -3709,7 +3710,10 @@ function commitPassiveMountOnFiber(
       }
 
       if (isViewTransitionEligible) {
-        if (supportsMutation && rootViewTransitionNameCanceled) {
+        if (
+          supportsViewTransition &&
+          rootViewTransitionNameCanceled
+        ) {
           restoreRootViewTransitionName(finishedRoot.containerInfo);
         }
       }
