+++ packages/react-reconciler/src/ReactFiberConfigWithNoMutation.js
@@ -37,25 +37,5 @@ export const hideTextInstance = shim;
 export const unhideInstance = shim;
 export const unhideTextInstance = shim;
 export const clearContainer = shim;
-export const applyViewTransitionName = shim;
-export const restoreViewTransitionName = shim;
-export const cancelViewTransitionName = shim;
-export const cancelRootViewTransitionName = shim;
-export const restoreRootViewTransitionName = shim;
-export const cloneRootViewTransitionContainer = shim;
-export const removeRootViewTransitionClone = shim;
-export type InstanceMeasurement = null;
-export const measureInstance = shim;
-export const measureClonedInstance = shim;
-export const wasInstanceInViewport = shim;
-export const hasInstanceChanged = shim;
-export const hasInstanceAffectedParent = shim;
-export const startViewTransition = shim;
-export type RunningViewTransition = null;
-export const startGestureTransition = shim;
-export const stopViewTransition = shim;
-export const addViewTransitionFinishedListener = shim;
-export type ViewTransitionInstance = null | {name: string, ...};
-export const createViewTransitionInstance = shim;
 export type GestureTimeline = any;
 export const getCurrentGestureOffset = shim;
